+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### 新特性
 * 【core   】     DynaBean.create增加重载方法（pr#245@Gitee）
+* 【core   】     IdcardUtil增加重载是否忽略大小写（issue#1348@Github）
 
 ### Bug修复